First test on Uno

+++ Hardware.cpp
@@ -40,14 +40,14 @@ void Hardware::setSignal(bool isMainTrack, bool high) {
 }
 
 int Hardware::getCurrentMilliamps(bool isMainTrack) {
-  int pin = isMainTrack ? MAIN_SENSE_PIN : PROG_SENSE_PIN;
+  byte  pin = isMainTrack ? MAIN_SENSE_PIN : PROG_SENSE_PIN;
   float factor = isMainTrack ? MAIN_SENSE_FACTOR : PROG_SENSE_FACTOR;
-  
+
   // IMPORTANT:  This function can be called in Interrupt() time within the 56uS timer
   //             The default analogRead takes ~100uS which is catastrphic
   //             so analogReadFast is used here. (-2uS) 
   int rawCurrent = analogReadFast(pin);
-  
+
   return (int)(rawCurrent * factor);
 }
